+++ package.json
@@ -23,6 +23,8 @@
   },
   "scripts": {
     "jsonlint": "jsonlint _data/simple-icons.json -q -V .jsonlintschema",
-    "svglint": "svglint icons/* --ci"
+    "svglint": "svglint icons/* --ci",
+    "prepublishOnly": "node scripts/prepublish.js",
+    "postpublish": "rm icons/*.js index.js"
   }
 }
